# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ai-agent/tools/companies/filter": {
      "post": {
        "operationId": "AiToolsController_companies",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BarChartQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "data: existing aggregation response; cacheHit: boolean"
          },
          "401": {
            "description": "Missing or invalid JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Read-only company aggregation through the existing DashboardService",
        "tags": [
          "AI Tools"
        ]
      }
    }
  },
  "info": {
    "title": "Dashboard API",
    "description": "Authentication, business management, dashboard aggregation, tenant-isolated AI records, conversations and read-only AI tools.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "YearRangeDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "number",
            "example": 1950,
            "minimum": 1800,
            "maximum": 2100
          },
          "end": {
            "type": "number",
            "example": 2023,
            "minimum": 1800,
            "maximum": 2100
          }
        }
      },
      "NumberRangeDto": {
        "type": "object",
        "properties": {
          "min": {
            "type": "number",
            "example": 0,
            "minimum": 0
          },
          "max": {
            "type": "number",
            "example": 1000000,
            "minimum": 0
          }
        }
      },
      "CompanyBarChartFilterDto": {
        "type": "object",
        "properties": {
          "level": {
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "country": {
            "example": [
              "China",
              "USA"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "city": {
            "example": [
              "Beijing",
              "New York"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "founded_year": {
            "$ref": "#/components/schemas/YearRangeDto"
          },
          "annual_revenue": {
            "$ref": "#/components/schemas/NumberRangeDto"
          },
          "employees": {
            "$ref": "#/components/schemas/NumberRangeDto"
          }
        }
      },
      "BarChartQueryDto": {
        "type": "object",
        "properties": {
          "dimension": {
            "type": "string",
            "enum": [
              "level",
              "country",
              "city"
            ],
            "example": "country"
          },
          "filter": {
            "$ref": "#/components/schemas/CompanyBarChartFilterDto"
          }
        },
        "required": [
          "dimension",
          "filter"
        ]
      },
...
    }
  }
}